Refactor packet processing (remove GPPacket)
